requirements.txt
pandas
numpy
openpyxl
PyYAML
tqdm
torch
faiss-cpu
sentence-transformers
fastapi
uvicorn
pydantic
